LayoutTests/accessibility/double-nested-inline-element-missing-from-tree-expected.txt

test1
test2
End of test
This tests that when you have a two nested inline continuations and the child one has no siblings, that we go back to the parent to check for its continuation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


AXRole: AXWebArea AXValue:
  AXRole: AXGroup AXValue:
    AXRole: AXStaticText AXValue: test1
    AXRole: AXGroup AXValue:
      AXRole: AXStaticText AXValue: test2
  AXRole: AXGroup AXValue:
PASS successfullyParsed is true

TEST COMPLETE
